+++ CommonCode/Machine/DefaultOps/Multiply.cs
@@ -6,7 +6,7 @@
     {
         public int DataLength => 3;
 
-        public void Act(IntMachine machine, ReadOnlySpan<int> opData, ReadOnlySpan<byte> mode)
+        public void Act(IntMachine machine, ReadOnlySpan<long> opData, ReadOnlySpan<byte> mode)
         {
             machine.Write(opData[2], machine.MarshallAccess(opData[0], mode[0]) * machine.MarshallAccess(opData[1], mode[1]));
         }
