+++ CommonCode/Machine/DefaultOps/OffsetRelativeBase.cs
@@ -6,7 +6,7 @@
     {
         public int DataLength => 1;
 
-        public void Act(IntMachine machine, ReadOnlySpan<int> opData, ReadOnlySpan<byte> modes)
+        public void Act(IntMachine machine, ReadOnlySpan<long> opData, ReadOnlySpan<byte> modes)
         {
             machine.OffsetRelativeBase(machine.MarshallAccess(opData[0], modes[0]));
         }
