+++ CommonCode/Machine/DefaultOps/Output.cs
@@ -6,7 +6,7 @@
     {
         public int DataLength => 1;
 
-        public void Act(IntMachine machine, ReadOnlySpan<int> opData, ReadOnlySpan<byte> mode)
+        public void Act(IntMachine machine, ReadOnlySpan<long> opData, ReadOnlySpan<byte> mode)
         {
             machine.SignalOutput(machine.MarshallAccess(opData[0], mode[0]));
         }
